#3 Switch alpine from latest tag to specific version
Following details in issue #3 , I would like to switch to using specific version rather than latest tag because docker images are not actually refreshed automatically which means last build is based on alpine:3.7 which is a bit outdated. Versioning with specific tag also allows to trigger build on version change or as @bsideup suggested monitoring it with dependabot.